Version 14_04_2021

+++ database.js
@@ -53,7 +53,7 @@ module.exports = {
 	
 	Add: async function(conn, id, Lora, Wifi, LTE, ip1, ip2){
 		let d = `Insert into mybd.Users(Id, LTE, LoRa, WiFi) values (${id}, ${LTE}, ${Lora}, ${Wifi})`;
-		console.log("id:"+id +" ip1:"+ ip1 +" ip2:"+ip2+" lora:"+Lora+" lte:"+LTE+" wifi:"+Wifi);
+		console.log("id: "+id +" ip1: "+ ip1 +" ip2: "+ip2+" lora: "+Lora+" lte: "+LTE+" wifi: "+Wifi);
 		await conn.execute(d);
 		if(Lora ==1){
 			let c = `Insert into mybd.lora(IdUser, IpIn, IpOut) values ('${id}', '${ip1}', '${ip2}')`;
@@ -70,13 +70,47 @@ module.exports = {
 	},
 	
 	Delete: async function(conn, id){
-		let d = `Delete FROM mybd.lte where IdUser= ${id}`;
-		await conn.execute(d);
-		let c = `Delete FROM mybd.wifi where IdUser= ${id}`;
-		await conn.execute(c);
-		let e = `Delete FROM mybd.lora where IdUser= ${id}`;
-		await conn.execute(e);
+		let arr = [];
+		let a = `SELECT * FROM mybd.users WHERE id=${id}`;
+		const [rows1,fields1] =  await conn.execute(a);
+		console.log(rows1);
+		if(rows1[0].LTE == 1){
+			let d1 = `SELECT * FROM mybd.lte WHERE IdUser=${id}`;
+			const [rows2,fields2] =  await conn.execute(d1);
+			console.log(rows2);
+			arr.push(""+rows2[0].IpIn);
+			arr.push(""+rows2[0].IpOut);
+			let d = `Delete FROM mybd.lte where IdUser= ${id}`;
+			await conn.execute(d);
+		}
+		if(rows1[0].LoRa==1){
+			if(rows1[0].LTE == 0){
+				let d1 = `SELECT * FROM mybd.lora WHERE IdUser=${id}`;
+				const [rows2,fields2] =  await conn.execute(d1);
+				console.log(rows2);
+				arr.push(""+rows2[0].IpIn);
+				arr.push(""+rows2[0].IpOut);
+			}
+		
+			let e = `Delete FROM mybd.lora where IdUser= ${id}`;
+			await conn.execute(e);
+			
+		}
+		if(rows1[0].WiFi==1)
+		{
+			if(rows1[0].LTE == 0 && rows1[0].LoRa == 0){
+				let d1 = `SELECT * FROM mybd.wifi WHERE IdUser=${id}`;
+				const [rows2,fields2] =  await conn.execute(d1);
+				console.log(rows2);
+				arr.push(""+rows2[0].IpIn);
+				arr.push(""+rows2[0].IpOut);
+			}
+			let c = `Delete FROM mybd.wifi where IdUser= ${id}`;
+			await conn.execute(c);
+		}
 		let k = `Delete FROM mybd.users where Id= ${id}`;
 		await conn.execute(k);
+		console.log(arr);
+		return arr;
 	}
 };
